+++ pulse_in.h
@@ -13,37 +13,37 @@ extern "C" {
 #endif
     void setup_pulse_in(void);
     
-    //Input channel 1 (int0)
+    //Pitch input (int0)
     //private
-    volatile long int0_start_time;
-    volatile long int0_width;
+    volatile unsigned int int0_start_time;
+    volatile unsigned int int0_width;
     
     void int0_isr(void);
-    long pitch_get_width(void);
+    unsigned int pitch_get_width(void);
 
-    //Input channel 2 (int1)
+    //Roll input (int1)
     //private
-    volatile long int1_start_time;
-    volatile long int1_width;
+    volatile unsigned int int1_start_time;
+    volatile unsigned int int1_width;
     
     void int1_isr(void);
-    long roll_get_width(void);
+    unsigned int roll_get_width(void);
 
-    //Input channel 3 (int2)
+    //Yaw input (int2)
     //private
-    volatile long int2_start_time;
-    volatile long int2_width;
+    volatile unsigned int int2_start_time;
+    volatile unsigned int int2_width;
     
     void int2_isr(void);
-    long yaw_get_width(void);
+    unsigned int yaw_get_width(void);
     
-    //Input channel 4 (cmr1)
+    //Throttle input (cmr1)
     //private
-    volatile long cmr1_start_time;
-    volatile long cmr1_width;
+    volatile unsigned int cmr1_start_time;
+    volatile unsigned int cmr1_width;
     
     void cmr1_isr(void);
-    long throttle_get_width(void);
+    unsigned int throttle_get_width(void);
     
 #ifdef	__cplusplus
 }
